saving output of training in files

+++ Main.py
@@ -243,8 +243,12 @@ def train_baseline(nets_, nets_path_, labeled_loader_: list, unlabeled_loader_, 
                     'SegNet3_Score': score_meters[2].value()[0].item(),
                     'MV_Score': ensemble_score.value()[0]}
         try:
-            pd.DataFrame([rec_data]).to_csv('baseline_03112018_segnet.csv',
-                                            index=False, float_format='%.4f')
+            if not os.path.exists():
+                pd.DataFrame([rec_data]).to_csv('baseline_03112018_segnet.csv',
+                                                index=False, float_format='%.4f')
+            else:
+                pd.DataFrame([rec_data]).to_csv('baseline_03112018_segnet.csv', mode='a',
+                                                index=False, float_format='%.4f')
         except Exception as e:
             print(e)
